feat: attendance-service 리팩토링 및 API 응답 일관성 강화
