+++ attendance-service/src/main/java/com/playdata/attendanceservice/absence/dto/response/AbsenceStatisticsDto.java
@@ -1,0 +1,17 @@
+package com.playdata.attendanceservice.absence.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class AbsenceStatisticsDto {
+    private long totalAbsences;
+    private long pendingAbsences;
+    private long approvedAbsences;
+    private long rejectedAbsences;
+}
