+++ attendance-service/src/main/java/com/playdata/attendanceservice/absence/entity/UrgencyType.java
@@ -1,0 +1,16 @@
+package com.playdata.attendanceservice.absence.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum UrgencyType {
+    URGENT("긴급"),
+    NORMAL("일반");
+
+    private final String description;
+
+    UrgencyType(String description) {
+        this.description = description;
+    }
+
+}
