--- attendance-service/src/main/java/com/playdata/attendanceservice/attendance/absence/dto/request/ApprovalRequestDto.java
@@ -1,24 +1,0 @@
-package com.playdata.attendanceservice.attendance.absence.dto.request;
-
-import lombok.*;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ApprovalRequestDto {
-    // setter 추가
-    @Setter
-    private Long absenceId;
-    private String requestType;
-    private Long applicantId;
-    private String applicantDepartment;
-    private String type;
-    private String urgency;
-    private String startDate;
-    private String endDate;
-    private String startTime;
-    private String endTime;
-    private String reason;
-
-}
