--- attendance-service/src/main/java/com/playdata/attendanceservice/attendance/absence/entity/UrgencyType.java
@@ -1,16 +1,0 @@
-package com.playdata.attendanceservice.attendance.absence.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum UrgencyType {
-    URGENT("긴급"),
-    NORMAL("일반");
-
-    private final String description;
-
-    UrgencyType(String description) {
-        this.description = description;
-    }
-
-}
