+++ attendance-service/src/main/java/com/playdata/attendanceservice/common/exception/AlreadyCheckedOutException.java
@@ -1,0 +1,11 @@
+package com.playdata.attendanceservice.common.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.CONFLICT) // 409 Conflict
+public class AlreadyCheckedOutException extends AttendanceException {
+    public AlreadyCheckedOutException(String message) {
+        super(message);
+    }
+}
