+++ attendance-service/src/main/java/com/playdata/attendanceservice/common/exception/CheckInNotFoundException.java
@@ -1,0 +1,11 @@
+package com.playdata.attendanceservice.common.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND) // 404 Not Found
+public class CheckInNotFoundException extends AttendanceException {
+    public CheckInNotFoundException(String message) {
+        super(message);
+    }
+}
