+++ attendance-service/src/main/resources/application.yml
@@ -5,6 +5,12 @@ spring:
   application:
     name: attendance-service
 
+  jpa:
+    show-sql: true
+    properties:
+      hibernate:
+        format_sql: true
+
 springdoc:
   api-docs:
     path: /v3/api-docs
@@ -23,11 +29,7 @@ standard:
     time: "09:00"
   checkout:
     time: "18:00"
-  jpa:
-    show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
+
 logging:
   level:
     org.hibernate.SQL: DEBUG
